MapReady merges

+++ prototype/src/osl/webservice.cpp
@@ -10,7 +10,7 @@
 osl::url_parser::url_parser(std::string URL) {
 	while (isspace(URL[0])) URL=URL.substr(1,std::string::npos);
 	unsigned int ds=URL.find_first_of("//"); /* double-slash */
-	if (ds!=std::string::npos) {
+	if (ds != (unsigned int) std::string::npos) {
 		protocol=URL.substr(0,ds-1);
 		ds+=2; /* skip over double-slash, so ds points at start of host */
 	} else { /* no double-slash: no protocol */
